rebasing onto nate's branch and moving remaining packages to org.bdgenomics
