+++ gnocchi-apis/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>net.fnothaft</groupId>
+    <groupId>org.bdgenomics</groupId>
     <artifactId>gnocchi-parent_2.10</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
@@ -71,7 +71,7 @@
       <artifactId>parquet-avro</artifactId>
     </dependency>
     <dependency>
-      <groupId>net.fnothaft</groupId>
+      <groupId>org.bdgenomics</groupId>
       <artifactId>gnocchi-core_2.10</artifactId>
       <version>0.0.1-SNAPSHOT</version>
     </dependency>
