+++ gnocchi-apis/src/test/java/org/bdgenomics/gnocchi/api/java/GnocchiFunSuite.scala
@@ -15,11 +15,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.fnothaft.gnocchi.assembly
+package org.bdgenomics.gnocchi.api.java
 
-/**
- * Empty assembly object, allows Maven build to create sources and javadoc artifacts.
- */
-object Assembly {
-  // empty
+import org.bdgenomics.utils.misc.SparkFunSuite
+
+trait GnocchiFunSuite extends SparkFunSuite {
+  override val appName: String = "gnocchi"
+  override val properties: Map[String, String] = Map(
+    "spark.serializer" -> "org.apache.spark.serializer.KryoSerializer",
+    "spark.kryo.registrator" -> "org.bdgenomics.adam.serialization.ADAMKryoRegistrator",
+    "spark.kryo.referenceTracking" -> "true",
+    "spark.driver.allowMultipleContexts" -> "false")
 }
+
