+++ gnocchi-assembly/pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>net.fnothaft</groupId>
+    <groupId>org.bdgenomics</groupId>
     <artifactId>gnocchi-parent_2.10</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
@@ -112,12 +112,12 @@
   </build>
   <dependencies>
     <dependency>
-      <groupId>net.fnothaft</groupId>
+      <groupId>org.bdgenomics</groupId>
       <artifactId>gnocchi-cli_2.10</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
-      <groupId>net.fnothaft</groupId>
+      <groupId>org.bdgenomics</groupId>
       <artifactId>gnocchi-apis.10</artifactId>
       <version>${project.version}</version>
     </dependency>
